+++ src/XRP/xrp-client-decorator.ts
@@ -5,6 +5,7 @@ import RawTransactionStatus from './raw-transaction-status'
 import XrpTransaction from './model/xrp-transaction'
 
 import SendXrpDetails from './model/send-xrp-details'
+import TransactionResult from './model/transaction-result'
 
 /** A decorator interface for XrpClients. */
 export default interface XrpClientDecorator {
@@ -115,4 +116,14 @@ export default interface XrpClientDecorator {
    * @returns An {@link XrpTransaction} object representing an XRP Ledger transaction.
    */
   getPayment(transactionHash: string): Promise<XrpTransaction | undefined>
+
+  /**
+   * Enable Deposit Authorization for this XRPL account.
+   * @see https://xrpl.org/depositauth.html
+   *
+   * @param wallet The wallet associated with the XRPL account enabling Deposit Authorization and that will sign the request.
+   * @returns A promise which resolves to a TransactionResult object that contains the hash of the submitted AccountSet transaction
+   *          and the final status of the transaction.
+   */
+  enableDepositAuth(wallet: Wallet): Promise<TransactionResult>
 }