+++ src/XRP/xrp-client.ts
@@ -8,6 +8,7 @@ import XrpClientInterface from './xrp-client-interface'
 import XrpTransaction from './model/xrp-transaction'
 
 import SendXrpDetails from './model/send-xrp-details'
+import TransactionResult from './model/transaction-result'
 
 /**
  * XrpClient is a client which interacts with the XRP Ledger.
@@ -138,5 +139,17 @@ export default class XrpClient implements XrpClientInterface {
     transactionHash: string,
   ): Promise<XrpTransaction | undefined> {
     return this.decoratedClient.getPayment(transactionHash)
+  }
+
+  /**
+   * Enable Deposit Authorization for this XRPL account.
+   * @see https://xrpl.org/depositauth.html
+   *
+   * @param wallet The wallet associated with the XRPL account enabling Deposit Authorization and that will sign the request.
+   * @returns A promise which resolves to a TransactionResult object that contains the hash of the submitted AccountSet transaction,
+   *          the final status of the transaction, and whether the transaction was included in a validated ledger.
+   */
+  enableDepositAuth(wallet: Wallet): Promise<TransactionResult> {
+    return this.decoratedClient.enableDepositAuth(wallet)
   }
 }