+++ test/Xpring/xpring-client.test.ts
@@ -9,6 +9,7 @@ import XpringClient from '../../src/Xpring/xpring-client'
 import RawTransactionStatus from '../../src/XRP/raw-transaction-status'
 
 import XpringError from '../../src/Xpring/xpring-error'
+import TransactionResult from '../../src/XRP/model/transaction-result'
 
 /* Default values for the fake XRP Client. These values must be provided but are not varied in testing. */
 /* eslint-disable @typescript-eslint/no-magic-numbers, @typescript-eslint/naming-convention --
@@ -28,6 +29,11 @@ const fakeRawTransactionStatus = new RawTransactionStatus(
 )
 const fakePaymentHistoryValue = []
 const fakeGetPaymentValue = testXrpTransaction
+const fakedEnableDepositAuthValue = new TransactionResult(
+  fakeTransactionHash,
+  TransactionStatus.Succeeded,
+  true,
+)
 /* eslint-enable @typescript-eslint/no-magic-numbers, @typescript-eslint/naming-convention */
 
 // An amount to send
@@ -56,6 +62,7 @@ describe('Xpring Client', function (): void {
       fakeAccountExistsResult,
       fakePaymentHistoryValue,
       fakeGetPaymentValue,
+      fakedEnableDepositAuthValue,
     )
 
     const resolvedXRPAddress = 'r123'
@@ -82,6 +89,7 @@ describe('Xpring Client', function (): void {
       fakeAccountExistsResult,
       fakePaymentHistoryValue,
       fakeGetPaymentValue,
+      fakedEnableDepositAuthValue,
     )
 
     const payIDClient = new FakeXrpPayIdClient(payIDError)
@@ -107,6 +115,7 @@ describe('Xpring Client', function (): void {
       fakeAccountExistsResult,
       fakePaymentHistoryValue,
       fakeGetPaymentValue,
+      fakedEnableDepositAuthValue,
     )
 
     const resolvedXRPAddress = 'r123'
@@ -133,6 +142,7 @@ describe('Xpring Client', function (): void {
       fakeAccountExistsResult,
       fakePaymentHistoryValue,
       fakeGetPaymentValue,
+      fakedEnableDepositAuthValue,
     )
 
     const payIDClient = new FakeXrpPayIdClient(payIDError)
@@ -158,6 +168,7 @@ describe('Xpring Client', function (): void {
       fakeAccountExistsResult,
       fakePaymentHistoryValue,
       fakeGetPaymentValue,
+      fakedEnableDepositAuthValue,
       XrplNetwork.Test,
     )
     const payIDClient = new FakeXrpPayIdClient(payIDError, XrplNetwork.Main)