+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+- A new method `enableDepositAuth` is added to `XrpClient` which enables Deposit Authorization for the specified XRPL account.
+(See https://xrpl.org/depositauth.html)
+
 ## 5.0.2 - 2020-08-01
 
 This release contains updated dependencies for stability and security.